requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
boto3==1.34.34
botocore==1.34.162
Brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
cssselect2==0.7.0
daphne==4.1.2
deep-translator==1.11.4
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.6
django-cors-headers==4.6.0
django-storages==1.14
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ecdsa==0.19.0
fonttools==4.56.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.8.2
jmespath==1.0.1
msgpack==1.1.0
oauthlib==3.2.2
openai==1.65.1
packaging==24.2
pdfkit==1.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.4
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
